+++ src/client/qcc/cli/__init__.py
@@ -7,10 +7,11 @@ from typing_extensions import Annotated
 
 from ..client import QCC
 from ..conf import settings
-from . import auth, session, static, status
+from . import auth, manage, session, static, status
 
 app = typer.Typer(name="qcc", no_args_is_help=True)
 app.add_typer(auth.app)
+app.add_typer(manage.app)
 app.add_typer(static.app)
 app.add_typer(status.app)
 app.add_typer(session.app)